Fix PillSheetModifiedHistoryAutomaticallyRecordedLastTakenDateAction
